requirement.txt
streamlit
gTTS
SpeechRecognition
google-generativeai
python-dotenv
PyMuPDF